fix home route to be same as index
